+++ test/models/messages/RemarksSpec.scala
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package models
+package models.messages
 
 import generators.{Generators, ModelGenerators}
 import models.XMLWrites._
-import models.messages._
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalatest.{FreeSpec, MustMatchers}
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
